move web socket logic out of the app
