+++ packages/front-end/src/App.css
@@ -1,30 +1,3 @@
-.app {
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.App {
   height: 100%;
-}
-
-.pokemon-draft-buttons {
-  display: flex;
-  margin-top: 34px;
-}
-
-.pokemon-draft-buttons button {
-  margin: 4px;
-  padding: 12px;
-}
-
-.app-pooled-pokemon-container {
-  width: 300px;
-  height: 100%;
-}
-
-.app-pokemon-information-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
 }
