+++ packages/front-end/src/components/PokemonLiveDraft.css
@@ -1,0 +1,30 @@
+.PokemonLiveDraft {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
+
+.PokemonLiveDraft-buttons {
+  display: flex;
+  margin-top: 34px;
+}
+
+.PokemonLiveDraft-buttons button {
+  margin: 4px;
+  padding: 12px;
+}
+
+.PokemonLiveDraft-pooled-pokemon-container {
+  width: 300px;
+  height: 100%;
+}
+
+.PokemonLiveDraft-pokemon-information-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
